git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@12159 842219ff-9341-0410-a943-573460bdd4de

+++ projects/CRM/docroot/includes/admintools_nav.jsp
@@ -1,4 +1,5 @@
 <%@ taglib uri='crm' prefix='crm' %>
+<%@ page import='com.freshdirect.webapp.crm.security.*' %>
 <%
 String snav_pageURI = request.getRequestURI();
 boolean resubmit_orders = snav_pageURI.indexOf("/admintools/index.jsp") > -1;
@@ -13,20 +14,46 @@ boolean top_faqs = snav_pageURI.indexOf("top_faqs") > -1;
 boolean click_to_call = snav_pageURI.indexOf("click_to_call") > -1;
 boolean search_orders = snav_pageURI.indexOf("search_orders_by_sku") > -1;
 boolean modify_orders = snav_pageURI.indexOf("modify_orders_by_sku") > -1;
+String lAgentId = CrmSecurityManager.getUserName(request);
+String lAgentRole = CrmSecurityManager.getUserRole(request);
 %>
-<crm:GetCurrentAgent id='currentAgent'>
+
 <div class="sup_nav_bg">
-<a href="/admintools/index.jsp" class="<%=resubmit_orders?"sup_nav_on":"sup_nav"%>">Resub. Orders</a>
-<a href="/admintools/resubmit_customers.jsp" class="<%=resubmit_customers?"sup_nav_on":"sup_nav"%>">Resub. Customers</a>
-<a href="/admintools/delete_reservations.jsp" class="<%=delete_reservations?"sup_nav_on":"sup_nav"%>">Del. Reservations</a>
-<a href="/admintools/mass_cancellation.jsp" class="<%=mass_cancellation?"sup_nav_on":"sup_nav"%>">Mass Cancellations</a>
-<a href="/admintools/mass_returns.jsp" class="<%=mass_returns?"sup_nav_on":"sup_nav"%>">Mass Returns</a>
-<a href="/admintools/delivery_restrictions.jsp" class="<%=delivery_restrictions?"sup_nav_on":"sup_nav"%>">Delivery Restrictions</a>
-<a href="/admintools/address_restrictions.jsp" class="<%=address_restrictions?"sup_nav_on":"sup_nav"%>">Address Restrictions</a>
-<a href="/admintools/settlement_batch.jsp?method=GET" class="<%=settlement_batch?"sup_nav_on":"sup_nav"%>">settlement batch</a>
-<a href="/admintools/top_faqs.jsp" class="<%=top_faqs?"sup_nav_on":"sup_nav"%>">Top 5 FAQ</a>
-<a href="/admintools/click_to_call.jsp" class="<%=click_to_call?"sup_nav_on":"sup_nav"%>">Click to Call Campaign</a>
-<a href="/admintools/search_orders_by_sku.jsp" class="<%=search_orders?"sup_nav_on":"sup_nav"%>">Search Orders By Skus</a>
-<a href="/admintools/modify_orders_by_sku.jsp?method=GET" class="<%=modify_orders?"sup_nav_on":"sup_nav"%>">Modify Orders By Skus</a>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"admintools_index.jsp")){ %>
+	<a href="/admintools/admintools_index.jsp" class="<%=resubmit_orders?"sup_nav_on":"sup_nav"%>">Resub. Orders</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"resubmit_customers.jsp")){ %>
+	<a href="/admintools/resubmit_customers.jsp" class="<%=resubmit_customers?"sup_nav_on":"sup_nav"%>">Resub. Customers</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"delete_reservations.jsp")){ %>
+	<a href="/admintools/delete_reservations.jsp" class="<%=delete_reservations?"sup_nav_on":"sup_nav"%>">Del. Reservations</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"mass_cancellation.jsp")){ %>
+	<a href="/admintools/mass_cancellation.jsp" class="<%=mass_cancellation?"sup_nav_on":"sup_nav"%>">Mass Cancellations</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"mass_returns.jsp")){ %>
+	<a href="/admintools/mass_returns.jsp" class="<%=mass_returns?"sup_nav_on":"sup_nav"%>">Mass Returns</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"delivery_restrictions.jsp")){ %>
+	<a href="/admintools/delivery_restrictions.jsp" class="<%=delivery_restrictions?"sup_nav_on":"sup_nav"%>">Delivery Restrictions</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"address_restrictions.jsp")){ %>
+	<a href="/admintools/address_restrictions.jsp" class="<%=address_restrictions?"sup_nav_on":"sup_nav"%>">Address Restrictions</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"settlement_batch.jsp")){ %>
+	<a href="/admintools/settlement_batch.jsp?method=GET" class="<%=settlement_batch?"sup_nav_on":"sup_nav"%>">settlement batch</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"top_faqs.jsp")){ %>
+	<a href="/admintools/top_faqs.jsp" class="<%=top_faqs?"sup_nav_on":"sup_nav"%>">Top 5 FAQ</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"click_to_call.jsp")){ %>
+	<a href="/admintools/click_to_call.jsp" class="<%=click_to_call?"sup_nav_on":"sup_nav"%>">Click to Call Campaign</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"search_orders_by_sku.jsp")){ %>
+	<a href="/admintools/search_orders_by_sku.jsp" class="<%=search_orders?"sup_nav_on":"sup_nav"%>">Search Orders By Skus</a>
+<% } %>
+<% if(CrmSecurityManager.hasAccessToPage(lAgentRole,"modify_orders_by_sku.jsp")){ %>
+	<a href="/admintools/modify_orders_by_sku.jsp?method=GET" class="<%=modify_orders?"sup_nav_on":"sup_nav"%>">Modify Orders By Skus</a>
+<% } %>
+
 </div>
-</crm:GetCurrentAgent>
